make order status component
