+++ src/components/store/index.ts
@@ -10,6 +10,7 @@ export { default as Hero } from './Hero';
 export { default as HiddenIcon } from './HiddenIcon';
 export { default as ItemCard } from './ItemCard';
 export { default as OrderCard } from './OrderCard';
+export { default as OrderStatusIndicator } from './OrderStatusIndicator';
 export { default as OrdersDisplay } from './OrdersDisplay';
 export { default as PickupEventDetail } from './PickupEventDetail';
 export { default as PickupEventPicker } from './PickupEventPicker';
